Create update_unattend.sh
DeployStudio script to move updated "unattend.xml" file onto an unbooted, SysPrepped BootCamp Windows7 partition. This is helpful if the unattend.xml on an image isn't working but you don't want to reSysPrep and retake an image!

The file to be copied down should be called "unattend.new" and reside in DeployStudio's "Files" folder.